Prefer current buffer when finding citekey heading
Several functions use `org-find-exact-heading-in-directory' to find the
location of a citekey heading. However, when called with a citekey at
point, it is likely that the citekey is in the current file (assuming
files are arranged by general topics), so first check the current buffer
and, only if it isn't found there, search all Org files in
`bog-notes-directory'.