Extend goto functions to support citekey properties
bog-goto-citekey-heading-in-buffer gives preference to citekeys as
heading titles over citekeys as properties.

The order of preference for bog-goto-citekey-heading-in-notes is the
citekey as a heading title in the current buffer, the citekey as a
property in the current buffer, the citekey as heading title in notes,
and, finally, the citekey as property in notes.