prepare-bib-file: Fix use of bibtex-key-in-head
Set the match data according to bibtex-entry-head before calling
bibtex-key-in-head.  bibtex-key-in-head relies on bibtex-entry-head
having generated the most recent match data, but I was using it as
though it took care of the matching.  By luck, it seems that the most
recent match data tends to be correct right after cleaning the entry.