Remove bog-goto-citekey-heading-in-buffer
This command wasn't too useful because, within a buffer, a citekey
heading can easily be jumped to with the standard org-goto interface.
Also, bog-goto-citekey-heading-in-notes will check the current file
first, so it behaves the same as bog-goto-citekey-heading-in-buffer,
except that it prompts with all citekeys instead of just those in the
current buffer.  Because a prompt limited to citekeys from the current
buffer can be useful, make this prompt available with a double C-u.