bog-citekey-tree-to-indirect-buffer: Use base-buffer
org-tree-to-indirect-buffer fails if it is called from the indirect
buffer create by the previous org-tree-to-indirect-buffer call.  Using
the base-buffer to find the citekey heading avoids this error.

Even if the error is fixed on Org's side, this change has the advantage
of preventing the name from appending the indirect buffer name rather
than the original note buffer's names.