travis: Update stale setup
The last Travis run didn't get out of the setup phase:

  https://travis-ci.org/github/kyleam/bog/builds/679436803

This updated .travis.yml is modified from Magit-annex's, which was
modified from Magit's.

Note that this goes from testing just Emacs 24 to testing the latest
Emacs release (26.3) and master.  We could of course do more---in
terms of both Emacs and Org versions---but, given the currently
dormant state of this project, two runs with different Emacs versions
and their built-in Orgs will do.